+++ src/command/index.ts
@@ -14,3 +14,4 @@ export { default as GeneratePreferenceCommand } from './GeneratePreferenceComman
 export { default as GenerateLayoutXmlCommand } from './GenerateLayoutXmlCommand';
 export { default as GenerateWebapiXmlFileCommand } from './GenerateWebapiXmlFileCommand';
 export { default as GeneratePageTypesXmlCommand } from './GeneratePageTypesXmlCommand';
+export { default as GenerateCrontabXmlCommand } from './GenerateCrontabXmlCommand';