+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added: Generator command for sample fieldset.xml file
 - Added: Generator command for sample view.xml file
 - Added: Generator command for sample indexer.xml file
+- Added: Generator command for sample mview.xml file
 
 ## [1.1.3] - 3/12/2025