+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added: Generator command for sample Layout XML file
 - Added: Generator command for sample page_types.xml file
 - Added: Generator command for sample crontab.xml file
+- Added: Generator command for sample email_templates.xml file
 
 ## [1.1.3] - 3/12/2025