+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added: Generator command for sample sections.xml file
 - Added: Generator command for sample fieldset.xml file
 - Added: Generator command for sample view.xml file
+- Added: Generator command for sample indexer.xml file
 
 ## [1.1.3] - 3/12/2025