+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added: Generator command for sample view.xml file
 - Added: Generator command for sample indexer.xml file
 - Added: Generator command for sample mview.xml file
+- Added: Generator command for sample widget.xml file
 
 ## [1.1.3] - 3/12/2025